# services/business-agency-saas-api/public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Laravel API Documentation'
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'You can retrieve your token by visiting your dashboard and clicking <b>Generate API token</b>.'
security:
  -
    default: []
paths:
...
  '/api/integrations/{id}':
    put:
      summary: ''
      operationId: putApiIntegrationsId
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                service:
                  type: string
                  description: ''
                  example: architecto
                value:
                  type: object
                  description: ''
                  example: []
                  properties: {  }
                key:
                  type: string
                  description: "Ensure 'value' is an array from frontend."
                  example: architecto
                  nullable: true
                is_active:
                  type: boolean
                  description: ''
                  example: false
                is_brain:
                  type: boolean
                  description: ''
                  example: true
              required:
                - service
                - value
                - is_active
                - is_brain
    delete:
      summary: 'DELETE /api/integrations/{id}'
      operationId: dELETEapiintegrationsid
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the integration.'
        example: architecto
        required: true
        schema:
          type: string
...